Add files via upload
Hello everyone, this is my first job, it's called the Game of life. The purpose of the program is to show how cells are born, live and die (in this case @). Although it is embellished, but I hope the essence is clear to everyone. Under certain conditions, the cell appears, lives and dies. In my code, a cell appears if it has 3 neighbors, lives if it has 2 or 3 neighbors, and dies if it has fewer or more than 2 or 3 neighbors.